feat(reporting): Phase 7 S8 — COREP C 02.00 as the recorded Kind-9 hybrid; the COREP credit-risk estate is fully declarative/delegated
The master own-funds roll-up moves to reporting/corep/c02.py per the
plan's §8.2 mandate: typed pre-pass aggregation kernels + a thin
row-assembly shell, deliberately NOT through the cellspec executor. The
one structural change: the generator's three instance-state dicts
(_irb_class_rwa / _slotting_type_rwa / _irb_sub_rwa, set-read-reset
across methods) become pure-function returns of _aggregate_by_approach;
17 functions relocated otherwise verbatim.

Preserved (recorded): col 0030 = the post-floor total (rwa_final is
ALREADY post-floor; rwa_pre_floor feeds only the 0034 activation boolean
with its 0.01 epsilon); equity RWA in three rows by design while the
flat total counts it once; SA rows keying RAW exposure_class through the
many-to-one accumulating class map; the _irb_*_split whole-total
fallbacks (corporate -> non-SME, RE -> residential-non-SME, retail-other
-> SME CRR-heritage); exposure_subclass as the canonical corporate split
with the flag heuristic fallback; the F-IRB/A-IRB FSE asymmetry; the B31
column policy (approach parents zero 0020/0030, sub-rows mirror 0010);
the 0500 currency-mismatch memo after the column pass (null comparison
columns, excluded from TREA); zero-fill vs null-fill row regimes; the
entity-scoped floor indicator rows.

Only C 34.x (CCR) remains imperative, deferred with S8-pre (operator).

Gate: goldens passed WITHOUT regeneration (both regimes); 64 C02 + 38
reporting-basis unit tests + 7 P2.41 subclass acceptance pins unchanged;
full suite 8,414 passed / 2 skipped / 5 xfailed; ruff, ty, arch_check
green; citation snapshot 158 fns.
