+++ scripts/arch_metrics.json
@@ -5,5 +5,5 @@
   "engine_collect_schema_sites": 154,
   "engine_eager_collect_sites": 33,
   "max_engine_module_loc": 1625,
-  "cites_decorators": 226
+  "cites_decorators": 227
 }
