+++ tests/contracts/data/citation_snapshot.json
@@ -452,6 +452,9 @@
   "rwa_calc.engine.supporting_factors::compute_e_star_group_drawn": [
     "CRR Art. 501"
   ],
+  "rwa_calc.reporting.corep.c02::generate_c02_00": [
+    "PS1/26, paragraph 1.3"
+  ],
   "rwa_calc.reporting.corep.c07::c07_population": [
     "PS1/26"
   ],
